Include mailer domain switch in devis mailers too
